added -h

+++ src/formatting/global.cpp
@@ -7,23 +7,23 @@
 #include "../headers/config.hpp"
 
 void removeSingleAnn(std::vector<std::string> &fileContents, std::string ann) {
-    for (size_t i = 0; i < fileContents.size(); i++) {
-        if (fileContents[i].find(ann) != (size_t)-1) {
+    for (int i = 0; i < fileContents.size(); i++) {
+        if (fileContents[i].find(ann) != (int)-1) {
             fileContents[i] = fileContents[i].substr(0, fileContents[i].find(ann));
         }
     }
 }
 void removeMultiAnn(std::vector<std::string> &fileContents, std::string ann) {
-    size_t pos;
-    size_t pos2;
-    for (size_t i = 0; i < fileContents.size(); i++) {
-        if ((pos=fileContents[i].find(ann)) != (size_t)-1) {
-            if ((pos2=fileContents[i].find(ann, pos + ann.length())) != (size_t)-1) {
+    int pos;
+    int pos2;
+    for (int i = 0; i < fileContents.size(); i++) {
+        if ((pos=fileContents[i].find(ann)) != (int)-1) {
+            if ((pos2=fileContents[i].find(ann, pos + ann.length())) != (int)-1) {
                 fileContents[i] = fileContents[i].substr(0, pos) + fileContents[i].substr(pos2 + ann.length());
             } else {
                 fileContents[i] = fileContents[i].substr(0, pos);
                 i++;
-                while ((pos=fileContents[i].find(ann)) == (size_t)-1) {
+                while ((pos=fileContents[i].find(ann)) == (int)-1) {
                     fileContents.erase(fileContents.begin() + i);
                     i++;
                 }
@@ -34,7 +34,7 @@ void removeMultiAnn(std::vector<std::string> &fileContents, std::string ann) {
 }
 
 void removeEmptyLines(std::vector<std::string> &fileContents) {
-    for (size_t i = 0; i < fileContents.size(); i++) {
+    for (int i = 0; i < fileContents.size(); i++) {
 		if (fileContents[i].empty()) {
 			fileContents.erase(fileContents.begin() + i);
 			i--;
@@ -43,15 +43,15 @@ void removeEmptyLines(std::vector<std::string> &fileContents) {
 }
 
 void replaceWinNewLines(std::vector<std::string> &fileContents) {
-    for (size_t i = 0; i < fileContents.size(); i++) {
+    for (int i = 0; i < fileContents.size(); i++) {
 		fileContents[i].erase(std::remove(fileContents[i].begin(), fileContents[i].end(), '\r'), fileContents[i].end());
 	}
 }
 
 void removeBlankLines(std::vector<std::string> &fileContents) {
     bool removeline = true;
-	for (size_t i = 0; i < fileContents.size(); i++) {
-		for (size_t j = 0; j < fileContents[i].length(); j++) if (fileContents[i][j] != ' ' && fileContents[i][j] != '\t') removeline = false;
+	for (int i = 0; i < fileContents.size(); i++) {
+		for (int j = 0; j < fileContents[i].length(); j++) if (fileContents[i][j] != ' ' && fileContents[i][j] != '\t') removeline = false;
 		if (removeline) {
 			fileContents.erase(fileContents.begin() + i);
 			i--;
@@ -63,8 +63,8 @@ void removeBlankLines(std::vector<std::string> &fileContents) {
 std::pair<bool,bool> isSpaced(std::vector<std::string> &fileContents) {
     bool isSpaced = false;
 	bool error = false;
-	for (size_t i = 0; i < fileContents.size(); i++) {
-		if (fileContents[i].find(":") != (size_t)-1) {
+	for (int i = 0; i < fileContents.size(); i++) {
+		if (fileContents[i].find(":") != (int)-1) {
 			if (fileContents[i+1][0] == ' ') isSpaced = true;
             else if (fileContents[i+1][0] != '\t') error = true;
 		}
@@ -78,9 +78,9 @@ std::tuple<std::vector<int>, int, int> detectIndentation(std::vector<std::string
     int spaces;
     std::vector<int> spacesVec;
     int indentation = 0;
-    for (size_t i = 0; i < fileContents.size(); i++) {
+    for (int i = 0; i < fileContents.size(); i++) {
         spaces = 0;
-        for (size_t j = 0; j < fileContents[i].length(); j++) {
+        for (int j = 0; j < fileContents[i].length(); j++) {
             if (fileContents[i][j] == ' ') while (fileContents[i][j] == ' ') {
                 spaces++;
                 j++;
@@ -89,7 +89,7 @@ std::tuple<std::vector<int>, int, int> detectIndentation(std::vector<std::string
         }
         spacesVec.push_back(spaces);
     }
-    for (size_t i = 0; i < spacesVec.size(); i++) {
+    for (int i = 0; i < spacesVec.size(); i++) {
         if (spacesVec[i] > 0) {
             for (int j = std::stoi(getValue("MAX_INDENT_SPACE")); j > 1; j--) {
                 if (spacesVec[i] % j == 0) {
@@ -100,12 +100,12 @@ std::tuple<std::vector<int>, int, int> detectIndentation(std::vector<std::string
             break;
         }
     }
-    for (size_t i = 0; i < spacesVec.size(); i++) if (spacesVec[i] % indentation != 0) {return std::tuple<std::vector<int>, int, int>(spacesVec, indentation, 1);}
+    for (int i = 0; i < spacesVec.size(); i++) if (spacesVec[i] % indentation != 0) {return std::tuple<std::vector<int>, int, int>(spacesVec, indentation, 1);}
     return std::tuple<std::vector<int>, int, int>(spacesVec, indentation, 0);
 }
 
 void replaceIndentation(std::vector<std::string> &fileContents, std::vector<int> &spacesVec, int indentation) {
-    for (size_t i = 0; i < fileContents.size(); i++) {
+    for (int i = 0; i < fileContents.size(); i++) {
         for (int j = spacesVec[i] / indentation; j > 0; j--) {
             fileContents[i].replace((indentation * j)-indentation, indentation, "\t");
         }
